Move MainGameFrame to main UI package, remove unused PlayerPanel.java, a duplicate of it exists which is used
